requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-ollama
chromadb
python-dotenv
atlassian-python-api
httpx
langchain-postgres
psycopg2-binary
langchain-google-genai
pydantic-settings
